+++ internal/namespace/definitons.go
@@ -7,13 +7,13 @@ import (
 
 type (
 	Namespace struct {
-		ID     int             `json:"id" db:"-" toml:"id"`
+		ID     int64           `json:"id" db:"-" toml:"id"`
 		Name   string          `json:"name" db:"-" toml:"name"`
 		Config json.RawMessage `json:"config,omitempty" db:"-" toml:"config,omitempty"`
 	}
 	Manager interface {
 		GetNamespaceByName(ctx context.Context, name string) (*Namespace, error)
-		GetNamespaceByConfigID(ctx context.Context, id int) (*Namespace, error)
+		GetNamespaceByConfigID(ctx context.Context, id int64) (*Namespace, error)
 		Namespaces(ctx context.Context) ([]*Namespace, error)
 	}
 	ManagerProvider interface {
